Modified all nodes depending on the go signal so that they now use UInt32 instead of Float32 for go message, should be more stable

+++ Drivers/motor/src/motor.h
@@ -27,11 +27,11 @@ unsigned int 	frontPWM,
 		rightPWM,
 		backPWM;
 
-float go = 0.0, go2 = 0.0, go3 = 0.0, go4 = 0.0;
+unsigned int go = 0, go2 = 0, go3 = 0, go4 = 0;
 
 /* Function Declarations */
 
-void goCallback(const std_msgs::Float32::ConstPtr& pilotGo);
+void goCallback(const std_msgs::UInt32::ConstPtr& pilotGo);
 void updatePWM(unsigned int channel, unsigned int rate);
 void frontCallback(const std_msgs::UInt32::ConstPtr& pidRampFront);
 void leftCallback(const std_msgs::UInt32::ConstPtr& pidRampLeft);
