+++ High_Level_Nodes/pilotKeys/src/pilotKey.h
@@ -4,7 +4,7 @@
 
 /* Globals*/
 
-float go=0.0;
+unsigned int go=0;
 float heading=0.0;
 float pitch=0.0;
 float depth=0.0;
@@ -12,7 +12,7 @@ float speed=0.0;
 
 /* Function Declerations */
 
-void goCallback(const std_msgs::Float32::ConstPtr& keyGo);
+void goCallback(const std_msgs::UInt32::ConstPtr& keyGo);
 void headingCallback(const std_msgs::Float32::ConstPtr& keyHeading);
 void pitchCallback(const std_msgs::Float32::ConstPtr& keyPitch);
 void depthCallback(const std_msgs::Float32::ConstPtr& keyDepth);
